Misc enhancements/bug fixes
Improved the magnitude and depth/time scales for all hypocenter plots

Changed name of color parameter to colorOption, to avoid intersection with the already in use color parameter for defining line color.  Renamed axes to axesOption for consistency with color option

Changed scale values for depth scales.  10km was changed to 13km for consistency with current analysis methods.

Added in auto scaling ability for depth scales.  Changed default value of hypocenter menu depth to defaults to 0 for min and auto for max

git-svn-id: http://avosouth.wr.usgs.gov/vhpsvn/Valve3/branches/isti_3.4@2106 b23e0666-5c4d-0410-a39a-a125d005d78a